backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
pyjwt==2.8.0
bcrypt==3.2.0
passlib[bcrypt]==1.7.4

# Validation & Settings
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Redis (for caching & sessions)
redis==5.0.1

# AWS S3 / File Upload
boto3==1.34.14
python-dotenv==1.0.0

# Razorpay
razorpay==1.4.1

# Utilities
httpx==0.26.0
python-dateutil==2.8.2
requests==2.31.0
cloudinary==1.36.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development
black==23.12.1
isort==5.13.2

